Add exception handler for watershed and conversion to uint8
The Vigra watershedsNew function supports a limited range of input
datatypes. If an input datatype was wrong, it would result in the
generic Vigra exception, which was unhelpful. This commit adds a
handler with a more descriptive message, and fixes a bug where
distance_transform_watershed passes the wrong datatype onwards.